[flink] Remove scan.push-down support which should be supported by Flink
