--- paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/source/table/BaseTableSource.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.flink.source.table;
-
-import org.apache.paimon.flink.source.FlinkTableSource;
-
-import org.apache.flink.table.connector.ChangelogMode;
-import org.apache.flink.table.connector.source.ScanTableSource;
-
-/** The {@link ScanTableSource} with push down interfaces. */
-public abstract class BaseTableSource implements ScanTableSource {
-
-    private final FlinkTableSource source;
-
-    public BaseTableSource(FlinkTableSource source) {
-        this.source = source;
-    }
-
-    @Override
-    public ChangelogMode getChangelogMode() {
-        return source.getChangelogMode();
-    }
-
-    @Override
-    public ScanRuntimeProvider getScanRuntimeProvider(ScanContext runtimeProviderContext) {
-        return source.getScanRuntimeProvider(runtimeProviderContext);
-    }
-
-    @Override
-    public String asSummaryString() {
-        return source.asSummaryString();
-    }
-}
