--- paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/source/table/PushedRichTableSource.java
@@ -1,65 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.flink.source.table;
-
-import org.apache.paimon.flink.source.FlinkTableSource;
-
-import org.apache.flink.table.connector.source.abilities.SupportsFilterPushDown;
-import org.apache.flink.table.connector.source.abilities.SupportsLimitPushDown;
-import org.apache.flink.table.connector.source.abilities.SupportsProjectionPushDown;
-import org.apache.flink.table.expressions.ResolvedExpression;
-
-import java.util.List;
-
-/** A {@link RichTableSource} with push down. */
-public class PushedRichTableSource extends RichTableSource
-        implements SupportsFilterPushDown, SupportsProjectionPushDown, SupportsLimitPushDown {
-
-    private final FlinkTableSource source;
-
-    public PushedRichTableSource(FlinkTableSource source) {
-        super(source);
-        this.source = source;
-    }
-
-    @Override
-    public PushedRichTableSource copy() {
-        return new PushedRichTableSource(source.copy());
-    }
-
-    @Override
-    public Result applyFilters(List<ResolvedExpression> filters) {
-        return Result.of(filters, source.pushFilters(filters));
-    }
-
-    @Override
-    public void applyLimit(long limit) {
-        source.pushLimit(limit);
-    }
-
-    @Override
-    public boolean supportsNestedProjection() {
-        return false;
-    }
-
-    @Override
-    public void applyProjection(int[][] projectedFields) {
-        source.pushProjection(projectedFields);
-    }
-}
