+++ paimon-flink/paimon-flink-common/src/test/java/org/apache/paimon/flink/source/FlinkTableSourceTest.java
@@ -52,13 +52,14 @@ public class FlinkTableSourceTest extends TableTestBase {
         Schema schema = Schema.newBuilder().column("col1", DataTypes.INT()).build();
         TableSchema tableSchema = new SchemaManager(fileIO, tablePath).createTable(schema);
         Table table = FileStoreTableFactory.create(LocalFileIO.create(), tablePath, tableSchema);
-        FlinkTableSource tableSource =
+        DataTableSource tableSource =
                 new DataTableSource(
                         ObjectIdentifier.of("catalog1", "db1", "T"), table, false, null, null);
 
         // col1 = 1
         List<ResolvedExpression> filters = ImmutableList.of(col1Equal1());
-        Assertions.assertThat(tableSource.pushFilters(filters)).isEqualTo(filters);
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
+                .isEqualTo(filters);
     }
 
     @Test
@@ -81,54 +82,57 @@ public class FlinkTableSourceTest extends TableTestBase {
 
         // col1 = 1 && p1 = 1 => [p1 = 1]
         List<ResolvedExpression> filters = ImmutableList.of(col1Equal1(), p1Equal1());
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .isEqualTo(ImmutableList.of(filters.get(0)));
 
         // col1 = 1 && p2 like '%a' => None
         filters = ImmutableList.of(col1Equal1(), p2Like("%a"));
-        Assertions.assertThat(tableSource.pushFilters(filters)).isEqualTo(filters);
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
+                .isEqualTo(filters);
 
         // col1 = 1 && p2 like 'a%' => [p2 like 'a%']
         filters = ImmutableList.of(col1Equal1(), p2Like("a%"));
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .isEqualTo(ImmutableList.of(filters.get(0)));
 
         // rand(42) > 0.1 => None
         filters = ImmutableList.of(rand());
-        Assertions.assertThat(tableSource.pushFilters(filters)).isEqualTo(filters);
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
+                .isEqualTo(filters);
 
         // upper(p1) = "A" => [upper(p1) = "A"]
         filters = ImmutableList.of(upperP2EqualA());
-        Assertions.assertThat(tableSource.pushFilters(filters)).isEqualTo(filters);
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
+                .isEqualTo(filters);
 
         // col1 = 1 && (p2 like 'a%' or p1 = 1) => [p2 like 'a%' or p1 = 1]
         filters = ImmutableList.of(col1Equal1(), or(p2Like("a%"), p1Equal1()));
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .isEqualTo(ImmutableList.of(filters.get(0)));
 
         // col1 = 1 && (p2 like '%a' or p1 = 1) => None
         filters = ImmutableList.of(col1Equal1(), or(p2Like("%a"), p1Equal1()));
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .containsExactlyInAnyOrder(filters.toArray(new ResolvedExpression[0]));
 
         // col1 = 1 && (p2 like 'a%' && p1 = 1) => [p2 like 'a%' && p1 = 1]
         filters = ImmutableList.of(col1Equal1(), and(p2Like("a%"), p1Equal1()));
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .isEqualTo(ImmutableList.of(filters.get(0)));
 
         // col1 = 1 && (p2 like '%a' && p1 = 1) => None
         filters = ImmutableList.of(col1Equal1(), and(p2Like("%a"), p1Equal1()));
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .containsExactlyInAnyOrder(filters.toArray(new ResolvedExpression[0]));
 
         // p2 like 'a%' && (col1 = 1 or p1 = 1) => [col1 = 1 or p1 = 1]
         filters = ImmutableList.of(p2Like("a%"), or(col1Equal1(), p1Equal1()));
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .isEqualTo(ImmutableList.of(filters.get(1)));
 
         // p2 like 'a%' && (col1 = 1 && p1 = 1) => [col1 = 1 && p1 = 1]
         filters = ImmutableList.of(p2Like("a%"), and(col1Equal1(), p1Equal1()));
-        Assertions.assertThat(tableSource.pushFilters(filters))
+        Assertions.assertThat(tableSource.applyFilters(filters).getRemainingFilters())
                 .isEqualTo(ImmutableList.of(filters.get(1)));
     }
